Added FPM script to build RPM
I added an FPM script, copied from other places where we use ./fpm.sh.
It will build a source RPM of the entire directory tracked by git.
Then it will run the `./configure` and `make` utilities to build the
object files in a temporary directory in the repo. Finally, it will
generate an RPM that installs the binaries from that temporary build
directory to the prefix "/usr/local".

This required adding a "VERSION" file to identify the version number
of the program.